flight-service/requirements.txt
beautifulsoup4
pandas
django
djangorestframework
django-cors-headers
requests
beautifulsoup4
django-prometheus